+++ src/write_outputs/write_outputs.jl
@@ -442,6 +442,13 @@ function write_outputs(EP::Model, path::AbstractString, setup::Dict, inputs::Dic
             println(elapsed_time_max_cap_req)
         end
 
+        if (setup["InvestmentIncentive"] == 1 || setup["ProductionIncentive"] == 1) &&
+           output_settings_d["WriteIncentives"]
+            elapsed_time_incentives = @elapsed write_incentives(path, inputs, setup, EP)
+            println("Time elapsed for writing incentives is")
+            println(elapsed_time_incentives)
+        end
+
         if setup["HydrogenMinimumProduction"] == 1 && has_duals(EP)
             if output_settings_d["WriteHydrogenPrices"]
                 elapsed_time_hydrogen_prices = @elapsed write_hydrogen_prices(path,